debug

+++ MP-SPDZ/Compiler/types.py
@@ -658,8 +658,10 @@ class _secret_structure(_structure):
                     else:
                         t = numpy.int64
                 if one_hot:
-                    content = numpy.eye(int(content.max()) + 1)[content]
-                content = content.astype(t)
+                    content = content.astype(int)
+content = numpy.eye(content.max() + 1)[content]
+
+content = content.astype(t)
                 f = program.get_binary_input_file(player)
                 f.write(content.tobytes())
                 f.flush()
